+++ crates/env/src/call/call_builder/call_v1.rs
@@ -24,9 +24,11 @@ use crate::{
         CallBuilder,
         ExecutionInput,
     },
-    Environment,
+    types::{
+        Environment,
+        Gas,
+    },
     Error,
-    Gas,
 };
 use num_traits::Zero;
 #[cfg(not(feature = "revive"))]