+++ crates/env/src/engine/mod.rs
@@ -81,7 +81,7 @@ pub(crate) fn decode_instantiate_result<I, E, ContractRef, R>(
 ) -> EnvResult<ConstructorResult<<R as ConstructorReturnType<ContractRef>>::Output>>
 where
     I: scale::Input,
-    E: crate::Environment,
+    E: crate::types::Environment,
     ContractRef: FromAccountId<E>,
     R: ConstructorReturnType<ContractRef>,
 {
@@ -106,7 +106,7 @@ fn decode_instantiate_err<I, E, ContractRef, R>(
 ) -> EnvResult<ConstructorResult<<R as ConstructorReturnType<ContractRef>>::Output>>
 where
     I: scale::Input,
-    E: crate::Environment,
+    E: crate::types::Environment,
     ContractRef: FromAccountId<E>,
     R: ConstructorReturnType<ContractRef>,
 {