+++ crates/primitives/src/lib.rs
@@ -28,8 +28,10 @@
 )]
 #![cfg_attr(not(feature = "std"), no_std)]
 
+mod arithmetic;
 mod key;
-mod types;
+pub mod reflect;
+pub mod types;
 
 pub use self::{
     key::{